requirements.txt
langchain
langchain-community
langchain-openai
chromadb==0.4.24
pypdf==5.0.0
sentence-transformers==3.0.0
requests==2.32.3
python-dotenv==1.1.0
tiktoken==0.7.0